+++ actionpack/lib/action_controller/metal/rendering.rb
@@ -37,7 +37,31 @@ module ActionController
     #   # => renders app/views/posts/show.html.erb
     #
     # If the first argument responds to +render_in+, the template will be
-    # rendered by calling +render_in+ with the current view context.
+    # rendered by calling +render_in+ with the current view context, render
+    # options, and block.
+    #
+    #   class Greeting
+    #     def render_in(view_context, **options, &block)
+    #       if block
+    #         view_context.render html: block.call
+    #       else
+    #         view_context.render inline: <<~ERB.strip, **options
+    #           <%= Hello, <%= local_assigns.fetch(:name, "World") %>
+    #         ERB
+    #       end
+    #     end
+    #
+    #     def format
+    #       :html
+    #     end
+    #   end
+    #
+    #   render(Greeting.new)                                        # => "Hello, World"
+    #   render(renderable: Greeting.new)                            # => "Hello, World"
+    #   render(Greeting.new, name: "Local")                         # => "Hello, Local"
+    #   render(renderable: Greeting.new, locals: { name: "Local" }) # => "Hello, Local"
+    #   render(Greeting.new) { "Hello, Block" }                     # => "Hello, Block"
+    #   render(renderable: Greeting.new) { "Hello, Block" }         # => "Hello, Block"
     #
     #   class Greeting
     #     def render_in(view_context)
@@ -110,11 +134,15 @@ module ActionController
     #
     # [+:renderable+]
     #   Renders the provided object by calling +render_in+ with the current view
-    #   context. The response format is determined by calling +format+ on the
-    #   renderable if it responds to +format+, falling back to +text/html+ by default.
+    #   context, render options, and block. The response format is determined by
+    #   calling +format+ on the renderable if it responds to +format+, falling
+    #   back to +text/html+ by default.
     #
     #     render renderable: Greeting.new
     #     # => renders "<h1>Hello, World</h1>"
+    #
+    #     render renderable: Greeting.new, locals: { name: "Local" }
+    #     # => renders "Hello, Local"
     #
     # By default, when a rendering mode is specified, no layout template is
     # rendered.